Added multiple things as the project was relatively simple and mostly for learning.
Created a seperate file for python functions from the main file.

This file contains functions to get the total word count, the total character count, and to sort the character count numerically.

Added a gitignore for ignoring a book directory that stores the data of books.

Added proper output formatting.